+++ app/gui2/src/components/GraphEditor/snap.ts
